ref: move forward command to core and URL helper to common
