+++ EMS/common-bundle/src/Helper/Url.php
@@ -2,8 +2,9 @@
 
 declare(strict_types=1);
 
-namespace App\CLI\Client\WebToElasticms\Helper;
+namespace EMS\CommonBundle\Helper;
 
+use App\CLI\Client\WebToElasticms\Helper\NotParsableUrlException;
 use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\PropertyInfo\Extractor\PhpDocExtractor;
 use Symfony\Component\PropertyInfo\Extractor\ReflectionExtractor;
