+++ EMS/core-bundle/src/Commands.php
@@ -42,4 +42,5 @@ final class Commands
     public const MANAGED_ALIAS_CREATE = 'ems:managed-alias:create';
     public const MANAGED_ALIAS_ADD_ENVIRONMENT = 'emsco:managed-alias:add-environment';
     public const ASSET_REFRESH_FILE_FIELD = 'emsco:asset:refresh-file-fields';
+    final public const SUBMISSION_FORWARD = 'emscli:form:forward';
 }
