+++ elasticms-cli/src/Client/Audit/Cache.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace App\CLI\Client\Audit;
 
-use App\CLI\Client\WebToElasticms\Helper\Url;
+use EMS\CommonBundle\Helper\Url;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\PropertyInfo\Extractor\PhpDocExtractor;
 use Symfony\Component\PropertyInfo\Extractor\ReflectionExtractor;
