+++ elasticms-cli/src/Client/WebToElasticms/Extract/Extractor.php
@@ -10,9 +10,9 @@ use App\CLI\Client\WebToElasticms\Config\ConfigManager;
 use App\CLI\Client\WebToElasticms\Config\Document as WebDocument;
 use App\CLI\Client\WebToElasticms\Config\WebResource;
 use App\CLI\Client\WebToElasticms\Helper\ExpressionData;
-use App\CLI\Client\WebToElasticms\Helper\Url;
 use App\CLI\Client\WebToElasticms\Rapport\Rapport;
 use EMS\CommonBundle\Elasticsearch\Document\Document;
+use EMS\CommonBundle\Helper\Url;
 use EMS\Helpers\Standard\Json;
 use GuzzleHttp\Exception\ClientException;
 use GuzzleHttp\Exception\RequestException;
