+++ elasticms-cli/src/Client/WebToElasticms/Filter/Attr/LinkMediaFile.php
@@ -5,8 +5,8 @@ declare(strict_types=1);
 namespace App\CLI\Client\WebToElasticms\Filter\Attr;
 
 use App\CLI\Client\WebToElasticms\Config\ConfigManager;
-use App\CLI\Client\WebToElasticms\Helper\Url;
 use App\CLI\Client\WebToElasticms\Rapport\Rapport;
+use EMS\CommonBundle\Helper\Url;
 
 class LinkMediaFile
 {
