+++ elasticms-cli/src/Command/Web/AuditCommand.php
@@ -11,7 +11,6 @@ use App\CLI\Client\Audit\Report;
 use App\CLI\Client\HttpClient\CacheManager;
 use App\CLI\Client\HttpClient\UrlReport;
 use App\CLI\Client\WebToElasticms\Helper\NotParsableUrlException;
-use App\CLI\Client\WebToElasticms\Helper\Url;
 use App\CLI\Commands;
 use Elastica\Query\BoolQuery;
 use Elastica\Query\Exists;
@@ -21,6 +20,7 @@ use EMS\CommonBundle\Common\Admin\AdminHelper;
 use EMS\CommonBundle\Common\Command\AbstractCommand;
 use EMS\CommonBundle\Contracts\CoreApi\Endpoint\Data\DataInterface;
 use EMS\CommonBundle\Elasticsearch\Document\EMSSource;
+use EMS\CommonBundle\Helper\Url;
 use EMS\Helpers\Standard\Json;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
